+++ base/exports.jl
@@ -953,9 +953,9 @@ export
     module_name,
     module_parent,
     names,
+    varinfo,
     versioninfo,
     which,
-    whos,
     workspace,
     @isdefined,